+++ packages/provider-utils/src/to-json-schema/effect-to-json-schema.ts
@@ -1,13 +1,12 @@
 import { JSONSchema7 } from '@ai-sdk/provider';
 
-export const effectToJsonSchema = async (
-  schema: unknown,
-): Promise<JSONSchema7> => {
-  try {
-    const { JSONSchema } = await import('effect');
-    return JSONSchema.make(schema as any) as JSONSchema7;
-  } catch {
-    // TODO dedicated error class
-    throw new Error(`Failed to import module 'effect'`);
-  }
-};
+export const effectToJsonSchema =
+  (schema: unknown) => async (): Promise<JSONSchema7> => {
+    try {
+      const { JSONSchema } = await import('effect');
+      return JSONSchema.make(schema as any) as JSONSchema7;
+    } catch {
+      // TODO dedicated error class
+      throw new Error(`Failed to import module 'effect'`);
+    }
+  };