+++ packages/huggingface/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @ai-sdk/huggingface
 
+## 1.0.0-beta.12
+
+### Patch Changes
+
+- Updated dependencies [b689220]
+  - @ai-sdk/openai-compatible@2.0.0-beta.17
+
+## 1.0.0-beta.11
+
+### Patch Changes
+
+- Updated dependencies [6306603]
+  - @ai-sdk/provider-utils@4.0.0-beta.16
+  - @ai-sdk/openai-compatible@2.0.0-beta.16
+
+## 1.0.0-beta.10
+
+### Patch Changes
+
+- Updated dependencies [f0b2157]
+  - @ai-sdk/provider-utils@4.0.0-beta.15
+  - @ai-sdk/openai-compatible@2.0.0-beta.15
+
 ## 1.0.0-beta.9
 
 ### Patch Changes